requirements.txt
streamlit
google-genai==0.8.0
python-dotenv
PyPDF2
streamlit-webrtc
openai